requirements.txt
DrissionPage>=4.0.0
dashscope>=1.14.0
requests>=2.31.0
pyyaml>=6.0.1
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0